+++ lib/transloadit/assembly.rb
@@ -173,8 +173,6 @@ class Transloadit::Assembly < Transloadit::ApiModel
     if is_retrying
       warn "Rate limit reached. Waiting for #{response.wait_time} seconds before retrying."
       sleep response.wait_time
-      # RestClient closes file streams at the end of a request.
-      ios.collect! { |file| File.open file.path }
     else
       raise Transloadit::Exception::RateLimitReached.new(response)
     end